Fastapi/requirements.txt
fastapi==0.68.0
pydantic~=1.8.2
uvicorn[standard]==0.13.0
pillow==8.3.2
python-multipart

